change license of s-express back to LGPL (with permission of all authors)
and explicitly add some licenses for copied code in other modules
