+++ project/EnsimeBuild.scala
@@ -93,10 +93,8 @@ object EnsimeBuild {
       libraryDependencies ++= Sensible.testLibs("compile")
     )
 
-  lazy val s_express = Project("s-express", file("s-express")) settings (commonSettings) dependsOn (
-    util,
-    testutil % "test"
-  ) settings (
+  lazy val s_express = Project("s-express", file("s-express")) settings (commonSettings) settings (
+      HeaderKey.headers := Copyright.LgplMap,
       libraryDependencies ++= Seq(
         "org.parboiled" %% "parboiled" % "2.1.2" // 2.1.3 doesn't have a _2.10
       ) ++ Sensible.shapeless(scalaVersion.value)
@@ -117,10 +115,9 @@ object EnsimeBuild {
 
   // the S-Exp protocol
   lazy val swanky = Project("swanky", file("protocol-swanky")) settings (commonSettings) dependsOn (
-    api,
+    api, s_express, util,
     testutil % "test",
-    api % "test->test", // for the test data
-    s_express
+    api % "test->test" // for the test data
   ) settings (
       libraryDependencies ++= Seq(
         "com.typesafe.akka" %% "akka-slf4j" % Sensible.akkaVersion
@@ -128,7 +125,7 @@ object EnsimeBuild {
     )
 
   lazy val core = Project("core", file("core")).dependsOn(
-    api, s_express, monkeys,
+    api, s_express, monkeys, util,
     api % "test->test", // for the interpolator
     testutil % "test,it",
     // depend on "it" dependencies in "test" or sbt adds them to the release deps!
