+++ s-express/src/main/scala/org/ensime/sexp/formats/FamilyFormats.scala
@@ -1,5 +1,5 @@
 // Copyright: 2010 - 2016 https://github.com/ensime/ensime-server/graphs
-// License: http://www.gnu.org/licenses/gpl-3.0.en.html
+// License: http://www.gnu.org/licenses/lgpl-3.0.en.html
 package org.ensime.sexp.formats
 
 import scala.collection.immutable.ListMap
