+++ s-express/src/test/scala/org/ensime/sexp/SexpParserSpec.scala
@@ -1,10 +1,8 @@
 // Copyright: 2010 - 2016 https://github.com/ensime/ensime-server/graphs
-// License: http://www.gnu.org/licenses/gpl-3.0.en.html
+// License: http://www.gnu.org/licenses/lgpl-3.0.en.html
 package org.ensime.sexp
 
-import org.ensime.util.EnsimeSpec
-
-class SexpParserSpec extends EnsimeSpec {
+class SexpParserSpec extends SexpSpec {
   import SexpParser.parse
 
   val foo = SexpString("foo")
