+++ s-express/src/test/scala/org/ensime/sexp/formats/BigIntConvertorSpec.scala
@@ -1,14 +1,15 @@
 // Copyright: 2010 - 2016 https://github.com/ensime/ensime-server/graphs
-// License: http://www.gnu.org/licenses/gpl-3.0.en.html
+// License: http://www.gnu.org/licenses/lgpl-3.0.en.html
 package org.ensime.sexp.formats
 
+import scala.collection.immutable.BitSet
+
 import BigIntConvertor._
+import org.ensime.sexp.SexpSpec
 import org.scalacheck.{ Arbitrary, Gen }
 import org.scalatest.prop.GeneratorDrivenPropertyChecks
-import scala.collection.immutable.BitSet
-import org.ensime.util.EnsimeSpec
 
-class BigIntConvertorSpec extends EnsimeSpec {
+class BigIntConvertorSpec extends SexpSpec {
   private val examples = List(
     BitSet() -> BigInt(0),
     BitSet(0) -> BigInt(1),
@@ -31,7 +32,7 @@ class BigIntConvertorSpec extends EnsimeSpec {
   }
 }
 
-class BigIntConvertorCheck extends EnsimeSpec with GeneratorDrivenPropertyChecks {
+class BigIntConvertorCheck extends SexpSpec with GeneratorDrivenPropertyChecks {
 
   def positiveIntStream: Arbitrary[Stream[Int]] = Arbitrary {
     Gen.containerOf[Stream, Int](Gen.chooseNum(0, 2 * Short.MaxValue))
