+++ s-express/src/test/scala/org/ensime/sexp/formats/FormatSpec.scala
@@ -1,12 +1,11 @@
 // Copyright: 2010 - 2016 https://github.com/ensime/ensime-server/graphs
-// License: http://www.gnu.org/licenses/gpl-3.0.en.html
+// License: http://www.gnu.org/licenses/lgpl-3.0.en.html
 package org.ensime.sexp.formats
 
 import org.ensime.sexp._
-import org.ensime.util.EnsimeSpec
 import org.scalactic.source.Position
 
-trait FormatSpec extends EnsimeSpec {
+trait FormatSpec extends SexpSpec {
   def assertFormat[T: SexpFormat](start: T, expect: Sexp)(implicit p: Position): Unit = {
     val sexp = start.toSexp
     assert(sexp === expect, s"${sexp.compactPrint} was not ${expect.compactPrint}")
